flash-fw: Add support to flash-fw
Introduce flash-fw.py script as a replacement of the current
flash-firmware.sh script.

New script is written in python and checks which are the released
firmwares available in share.loverpi.com. It also checks for the
available boards.
Finally it can start the flashing firmware procedure (boot.py) by
downloading the released binary for the given board. For that, it first
checks the board is connected through USB, downloads the selected
firwmare version (or uses the local if provided) and starts the
flashing.

Signed-off-by: Daniel Gomez <dagmcr@gmail.com>